updates on source code

+++ database/migrations/2024_10_29_064711_create_hris_employee_leave_balances_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('emp_id');
             $table->integer('leave_type_id');
-            $table->integer('leave_balance');
+            $table->float('leave_balance');
 
             $table->tinyInteger('is_active')->nullable();
             $table->tinyInteger('is_deleted')->nullable();
